Groups edit implemented

+++ app/Http/Controllers/GroupsController.php
@@ -13,29 +13,63 @@ class GroupsController extends Controller
         $this->middleware('verified');
     }
 
-    public function index($code){
-       
+    public function index($code)
+    {
         $class = Classroom::where('code', '=', $code)->with('grouping.groups', 'students.groups')->firstorFail();
         $this->authorize('view', $class);
         return view('groups.index', compact('class'));
-        
     }
 
-    public function update($code) {
+    public function update($code, $type)
+    {
         $class = Classroom::where('code', '=', $code)->firstorFail();
         $this->authorize('update', $class);
-        foreach (request()->dropGroups as $group) {
-            $groupObj = Group::findOrFail($group['id']);
-            if($groupObj->grouping->classroom_id != $class->id)
-                continue;
-            $groupObj->students()->detach($groupObj->students);
-            foreach ($group['children'] as $std) {
-                $groupObj->students()->attach($std['id']);
+
+        if ($type == 'students') {
+            foreach (request()->dropGroups as $group) {
+                $groupObj = Group::findOrFail($group['id']);
+                if ($groupObj->grouping->classroom_id != $class->id)
+                    continue;
+                $groupObj->students()->detach($groupObj->students);
+                foreach ($group['children'] as $std) {
+                    $groupObj->students()->attach($std['id']);
+                }
             }
+        } else if ($type == 'info') {
+            $data = request()->validate([
+                'id' => ['numeric'],
+                'name' => ['string'],
+                'logo' => ['image', 'nullable'],
+            ]);
+
+            $group = Group::find($data['id']);
+            $class = Classroom::where('id', '=', $group->grouping->classroom_id)->firstOrFail();
+            $this->authorize('update', $class);
+
+            if (request()->logo) {
+                $media = $group->addMedia(request()->file('logo'))
+                    ->toMediaCollection('logo');
+                dump($media->getUrl());
+                $group->update(['logo' => $media->getUrl()]);
+            } 
+            $group->update(['name' => $data['name']]);
         }
     }
-    
-    public function store($code) {
+
+    public function destroy($id) {
+        $group = Group::find($id);
+        dump($id);
+        $this->authorize('update', Classroom::where('id', $group->grouping->classroom_id)->firstOrFail());
+        try {
+            $group->delete();
+        } catch (\Throwable $th) {
+            return [ 'error' => $th];
+        }
+        return 1;
+    }
+
+    public function store($code)
+    {
         $class = Classroom::where('code', '=', $code)->firstorFail();
         $this->authorize('update', $class);
         $data = request()->validate([
@@ -45,16 +79,13 @@ class GroupsController extends Controller
         $generator = new \Nubs\RandomNameGenerator\Alliteration();
 
         $groups = [];
-        for ($i=0; $i < $data['groupsNumber']; $i++) {
+        for ($i = 0; $i < $data['groupsNumber']; $i++) {
 
             $groups[] = Group::create([
                 'name' => $generator->getName(),
                 'grouping_id' => $class->grouping->first()->id,
             ]);
-
         }
         return $groups;
-
-                
     }
 }
